+++ src/main/java/com/clinica/agendamento_consulta_medica/service/SpecialtyService.java
@@ -1,7 +1,12 @@
 package com.clinica.agendamento_consulta_medica.service;
 import com.clinica.agendamento_consulta_medica.dto.specialty.SpecialtyRequestDTO;
+import com.clinica.agendamento_consulta_medica.dto.specialty.SpecialtyResponseDTO;
+import com.clinica.agendamento_consulta_medica.entities.Doctor;
 import com.clinica.agendamento_consulta_medica.entities.Specialty;
+import com.clinica.agendamento_consulta_medica.repository.DoctorRepository;
 import com.clinica.agendamento_consulta_medica.repository.SpecialtyRepository;
+import com.clinica.agendamento_consulta_medica.service.exception.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.stereotype.Service;
 import java.util.List;
 import java.util.Optional;
@@ -10,46 +15,63 @@ import java.util.stream.Collectors;
 @Service
 public class SpecialtyService {
 
-    private  final SpecialtyRepository especialtyRepository;
+    private final SpecialtyRepository especialtyRepository;
+    private final DoctorRepository doctorRepository;
 
-    public SpecialtyService(SpecialtyRepository especialtyRepository) {
+    public SpecialtyService(SpecialtyRepository especialtyRepository, DoctorRepository doctorRepository) {
         this.especialtyRepository = especialtyRepository;
+        this.doctorRepository = doctorRepository;
     }
 
-    public SpecialtyRequestDTO save (SpecialtyRequestDTO especialty){
+    public SpecialtyResponseDTO save(SpecialtyRequestDTO specialtyRequestDTO) {
+
+        Doctor doctor = doctorRepository.findById(specialtyRequestDTO.getDoctorId()).orElseThrow(() ->
+                new ResourceNotFoundException(specialtyRequestDTO.getDoctorId()));
+
         Specialty specialty = new Specialty();
-        specialty.setId(especialty.getId());
-        specialty.setName(especialty.getName());
-        specialty.setPrice(especialty.getPrice());
+
+        specialty.setName(specialtyRequestDTO.getName());
+        specialty.setPrice(specialtyRequestDTO.getPrice());
+        specialty.setDoctors(doctor);
         especialtyRepository.save(specialty);
-        return new SpecialtyRequestDTO(specialty);
+        return new SpecialtyResponseDTO(specialty);
     }
 
-    public List<SpecialtyRequestDTO> findAll(){
-        List<Specialty> specialty =  especialtyRepository.findAll();
-        return specialty.stream().map(SpecialtyRequestDTO::new).collect(Collectors.toList());
+    public List<SpecialtyResponseDTO> findAll() {
+        List<Specialty> specialty = especialtyRepository.findAll();
+        return specialty.stream().map(SpecialtyResponseDTO::new).collect(Collectors.toList());
     }
 
-    public SpecialtyRequestDTO findById(Long id){
+    public SpecialtyResponseDTO findById(Long id) {
         Optional<Specialty> especialty = especialtyRepository.findById(id);
-        return new SpecialtyRequestDTO(especialty.get());
+        return new SpecialtyResponseDTO(especialty.orElseThrow(() -> new ResourceNotFoundException(id)));
     }
 
-    public void deleteById(Long id){
+    public void deleteById(Long id) {
+        if (!especialtyRepository.existsById(id)) {
+            throw new ResourceNotFoundException(id);
+        }
         especialtyRepository.deleteById(id);
     }
 
+    public SpecialtyResponseDTO update(Long id, SpecialtyRequestDTO especialty) {
 
-    public SpecialtyRequestDTO update(Long id, SpecialtyRequestDTO especialty){
-        Specialty especialty1 = especialtyRepository.getReferenceById(id);
-        updateData(especialty1,especialty);
-         especialtyRepository.save(especialty1);
-         return new SpecialtyRequestDTO(especialty1);
+        try {
+            Specialty especialty1 = especialtyRepository.getReferenceById(id);
+            updateData(especialty1, especialty);
+            especialtyRepository.save(especialty1);
+            return new SpecialtyResponseDTO(especialty1);
+        } catch (EntityNotFoundException exception) {
+            throw new ResourceNotFoundException(id);
+        }
     }
 
-    private void updateData(Specialty especialty1, SpecialtyRequestDTO especialty) {
-            especialty1.setName(especialty.getName());
-            especialty1.setPrice(especialty.getPrice());
+    private void updateData(Specialty especialty1, SpecialtyRequestDTO specialtyRequestDTO) {
+        Doctor doctor = doctorRepository.findById(specialtyRequestDTO.getDoctorId()).orElseThrow(() ->
+                new ResourceNotFoundException(specialtyRequestDTO.getDoctorId()));
 
+        especialty1.setName(specialtyRequestDTO.getName());
+        especialty1.setDoctors(doctor);
+        especialty1.setPrice(specialtyRequestDTO.getPrice());
     }
 }
